requirements.txt
streamlit
groq
python-dotenv
# add any other libraries you imported at the top of your script